refactor: moved components to shared folder for compatibility
